Split also a PS test.
