+++ tests/CMakeLists.txt
@@ -58,6 +58,7 @@ ADD_PIRANHA_TESTCASE(parallel_vector_transform)
 ADD_PIRANHA_TESTCASE(poisson_series_01)
 ADD_PIRANHA_TESTCASE(poisson_series_02)
 ADD_PIRANHA_TESTCASE(poisson_series_03)
+ADD_PIRANHA_TESTCASE(poisson_series_04)
 ADD_PIRANHA_TESTCASE(polynomial_01)
 ADD_PIRANHA_TESTCASE(polynomial_02)
 ADD_PIRANHA_TESTCASE(polynomial_03)
